# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym Project API",
    "version": "1.0.0",
    "description": "API для фитнес-трекера"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/exercises/": {
      "get": {
        "summary": "Список упражнений",
        "tags": ["Exercises"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "muscle_group",
            "in": "query",
            "description": "Фильтр по ID группы мышц",
            "schema": { "type": "integer" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Поиск по названию и описанию",
            "schema": { "type": "string" }
          },
          {
            "name": "ordering",
            "in": "query",
            "description": "Сортировка (name, id, -name, -id)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Exercise" }
                }
              }
            }
          },
          "401": { "description": "Не авторизован" }
        }
      },
      "post": {
        "summary": "Создать упражнение (только admin)",
        "tags": ["Exercises"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExerciseWrite" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Exercise" }
              }
            }
          },
          "400": { "description": "Ошибка валидации" },
          "403": { "description": "Нет прав (не admin)" }
        }
      }
    },
...
    "/workouts/create/": {
      "post": {
        "summary": "Создать тренировку на дату",
        "description": "Создаёт тренировку на указанную дату. Если тренировка уже существует — возвращает 409.",
        "tags": ["Workouts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-06-01"
                  }
                },
                "required": ["date"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Workout" }
              }
            }
          },
          "400": { "description": "Неверный формат даты" },
          "401": { "description": "Не авторизован" },
          "409": { "description": "Тренировка на эту дату уже существует" }
        }
      }
    },
...
    "/workouts/today/": {
      "get": {
        "summary": "Сегодняшняя тренировка (автосоздание)",
        "tags": ["Workouts"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Workout" }
              }
            }
          }
        }
      }
    },
...
    "/workouts/{workout_id}/stats/": {
      "get": {
        "summary": "Статистика тренировки",
        "tags": ["Workouts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/workout_id" }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkoutStats" }
              }
            }
          }
        }
      }
    },
    "/workouts/{workout_id}/exercises/": {
      "post": {
        "summary": "Добавить упражнение в тренировку",
        "tags": ["Workouts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/workout_id" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "exercise_id": { "type": "integer", "example": 1 },
                  "exercise_order": { "type": "integer", "example": 1 }
                },
                "required": ["exercise_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Добавлено",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkoutExercise" }
              }
            }
          },
          "409": { "description": "Уже в тренировке" }
        }
      }
    },
    "/workouts/{workout_id}/exercises/{exercise_id}/": {
      "delete": {
        "summary": "Удалить упражнение из тренировки",
        "tags": ["Workouts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/workout_id" },
          { "$ref": "#/components/parameters/exercise_id" }
        ],
        "responses": {
          "204": { "description": "Удалено" }
        }
      }
    },
    "/workouts/{workout_id}/exercises/{exercise_id}/sets/": {
      "post": {
        "summary": "Добавить подход",
        "tags": ["Sets"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/workout_id" },
          { "$ref": "#/components/parameters/exercise_id" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "weight": { "type": "number", "example": 80 },
                  "reps": { "type": "integer", "example": 10 },
                  "set_number": { "type": "integer", "example": 1 }
                },
                "required": ["weight", "reps"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Создано",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Set" }
              }
            }
          },
          "400": { "description": "Вес ≤ 0 или повторения ≤ 0" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "workout_id": {
        "name": "workout_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "exercise_id": {
        "name": "exercise_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
...
    },
    "schemas": {
...
      "Exercise": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "muscle_group": { "type": "integer", "description": "ID группы мышц" },
          "muscle_group_name": { "type": "string", "readOnly": true },
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "default": "" },
          "image": { "type": "string", "format": "uri", "nullable": true, "readOnly": true }
        },
        "required": ["muscle_group", "muscle_group_name", "name"]
      },
      "ExerciseWrite": {
        "type": "object",
        "properties": {
          "muscle_group": { "type": "integer", "description": "ID группы мышц" },
          "name": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "default": "" },
          "image": { "type": "string", "format": "binary", "nullable": true }
        },
        "required": ["muscle_group", "name"]
      },
...
      "Workout": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "date": { "type": "string", "format": "date" },
          "exercises": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WorkoutExercise" }
          },
          "exercise_count": { "type": "integer" },
          "set_count": { "type": "integer" },
          "total_weight": { "type": "number" }
        }
      },
      "WorkoutExercise": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "exercise_id": { "type": "integer", "writeOnly": true },
          "exercise_name": { "type": "string", "readOnly": true },
          "muscle_group": { "type": "string", "readOnly": true },
          "exercise_order": { "type": "integer" },
          "sets": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Set" }
          }
        }
      },
      "Set": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "weight": { "type": "number", "example": 80 },
          "reps": { "type": "integer", "example": 10 },
          "set_number": { "type": "integer", "example": 1 }
        }
      },
      "WorkoutStats": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "exercise_count": { "type": "integer" },
          "set_count": { "type": "integer" },
          "total_weight": { "type": "number" }
        }
      },
...
    }
  }
}